v1
You can add new enemies by changing the characters in the config file. The enemy character that comes after the "Character" in the expression "Character:Darthvader,Door:B" is the name of the character and can only be Darthvader, Stormtrooper, Kyloren. Darthvader can break walls. Kyloren can move 2 units and stormtrooper can move 1 unit. The expression "Door" indicates the doors where enemy characters will spawn. These are shown with arrows on the map. So you can add as many enemies as you want to these doors. I set the config file as default to make the game difficult so that you can win the game with only 1 few chances. Also, the paths expressed with "1" in the matrix in the config file. "0" indicates the walls. You can change these.